backend/trajectory-predictor/requirements.txt
torch
numpy
einops
pika
pymysql
cryptography
python-dotenv
